requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.55.0
anyio==4.9.0
asyncer==0.0.7
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
blis==1.3.0
build==1.2.2.post1
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.6.15
chainlit==2.5.5
charset-normalizer==3.4.2
chevron==0.14.0
chromadb==1.0.13
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
cymem==2.0.11
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
durationpy==0.10
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.115.14
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.3
groq==0.29.0
grpcio==1.73.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflection==0.5.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kubernetes==33.1.0
langchain==0.3.26
langchain-chroma==0.2.4
langchain-community==0.3.26
langchain-core==0.3.66
langchain-groq==0.3.4
langchain-huggingface==0.3.0
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
langcodes==3.5.0
langsmith==0.4.4
language_data==1.3.0
Lazify==0.4.0
literalai==0.1.201
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.10.1
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.6.2
murmurhash==1.0.13
mypy_extensions==1.1.0
narwhals==1.44.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.1
oauthlib==3.3.1
onnxruntime==1.22.0
openai==1.93.0
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-alephalpha==0.40.14
opentelemetry-instrumentation-anthropic==0.40.14
opentelemetry-instrumentation-bedrock==0.40.14
opentelemetry-instrumentation-chromadb==0.40.14
opentelemetry-instrumentation-cohere==0.40.14
opentelemetry-instrumentation-crewai==0.40.14
opentelemetry-instrumentation-google-generativeai==0.40.14
opentelemetry-instrumentation-groq==0.40.14
opentelemetry-instrumentation-haystack==0.40.14
opentelemetry-instrumentation-lancedb==0.40.14
opentelemetry-instrumentation-langchain==0.40.14
opentelemetry-instrumentation-llamaindex==0.40.14
opentelemetry-instrumentation-logging==0.55b1
opentelemetry-instrumentation-marqo==0.40.14
opentelemetry-instrumentation-mcp==0.40.14
opentelemetry-instrumentation-milvus==0.40.14
opentelemetry-instrumentation-mistralai==0.40.14
opentelemetry-instrumentation-ollama==0.40.14
opentelemetry-instrumentation-openai==0.40.14
opentelemetry-instrumentation-pinecone==0.40.14
opentelemetry-instrumentation-qdrant==0.40.14
opentelemetry-instrumentation-redis==0.55b1
opentelemetry-instrumentation-replicate==0.40.14
opentelemetry-instrumentation-requests==0.55b1
opentelemetry-instrumentation-sagemaker==0.40.14
opentelemetry-instrumentation-sqlalchemy==0.55b1
opentelemetry-instrumentation-threading==0.55b1
opentelemetry-instrumentation-together==0.40.14
opentelemetry-instrumentation-transformers==0.40.14
opentelemetry-instrumentation-urllib3==0.55b1
opentelemetry-instrumentation-vertexai==0.40.14
opentelemetry-instrumentation-watsonx==0.40.14
opentelemetry-instrumentation-weaviate==0.40.14
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-semantic-conventions-ai==0.4.9
opentelemetry-util-http==0.55b1
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.3.0
pillow==11.2.1
posthog==3.25.0
preshed==3.0.10
propcache==0.3.2
protobuf==5.29.5
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.1
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-multipart==0.0.18
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
sentence-transformers==4.1.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
sse-starlette==2.3.6
starlette==0.41.3
sympy==1.14.0
syncer==2.0.3
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
toml==0.10.2
tomli==2.2.1
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
traceloop-sdk==0.40.14
transformers==4.53.0
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uptrace==1.34.0
urllib3==2.5.0
uvicorn==0.35.0
wasabi==1.1.3
watchdog==6.0.0
watchfiles==0.20.0
weasel==0.4.1
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
